Move RemoteCryptoMarketLoader to a file
